requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
phonenumbers~=8.13.22
twilio~=8.9.1
decouple~=0.0.7